+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Changes
 * Added ability to hide metrics for Elasticsearch MetricAggs
 * Add derivative metric aggregation for Elasticsearch
 * Add ``Stat`` class (and ``StatMapping``, ``StatValueMapping``, ``StatRangeMapping``) to support the Stat panel
+* Add ``Svg`` class to support the SVG panel
 * ...
 
 
@@ -37,6 +38,7 @@ Changes
 * Add ``Heatmap`` class (and ``HeatmapColor``) to support the Heatmap panel (#170)
 * Add ``BarGuage`` for creating bar guages panels in grafana 6
 * Add ``GuagePanel`` for creating guages in grafana 6
+* Add data links support to ``Graph``, ``BarGuage``, and ``GuagePanel`` panels
 * Removed gfdatasource - feature is built in to Grafana since v5.
 * Generate API docs for readthedocs.org
 * Fix AlertList panel generation